Allow ns_lists of incomplete types
Allow declaration of a list head whose element structure type is
incomplete.

This permits better abstraction in header files - with the previous
setup all structures had to be fully visible, but you can now declare
a list of a structure that you do not have access to.

The restriction is that the link element must be the first element in
the complete structure.

Link offset type storage modified to actually permit the link element
to be the first element - it failed if the offset was 0, due to
declaring a zero-length array.