libService: add new function to nsdynmem to query the mem stat structure
Add new call to query the memory statistics structure which is
given during the ns_dyn_mem_init() phase. This is useful for reporting
the memory usage statistics, as the heap init phase is board/platform
spesific and the stats structure given at that time may not be visible
to the testing code side.